Added A StartScreen/Mode Selection/ Basic Level Advancement/ Minor Bugfixes

+++ Minigin/Font.cpp
@@ -8,7 +8,8 @@ TTF_Font* Font::GetFont() const {
 
 Font::Font(const std::string& fullPath, unsigned int size) : m_font(nullptr), m_size(size)
 {
-	m_font = TTF_OpenFont(fullPath.c_str(), size);
+	m_font = TTF_OpenFont(fullPath.c_str(), static_cast<signed int>(size));
+
 	if (m_font == nullptr)
 	{
 		throw std::runtime_error(std::string("Failed to load font: ") + SDL_GetError());
